+++ modules/purchase/__init__.py
@@ -1,0 +1,1 @@
+"""Módulo de compras (diseño F5-01)."""